agents/test-agent/requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0
boto3>=1.34.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
